+++ Tp-Backend/src/main/java/com/example/tpbackend/repository/utilisateur/GestionnaireRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface GestionnaireRepository extends JpaRepository<Gestionnaire, Integer> {
+public interface GestionnaireRepository extends JpaRepository<Gestionnaire, Long> {
 
     boolean existsByMatricule(String matricule);
